Move classes to the default packages
