+++ awslambda/src/test/scala/docs/scaladsl/AwsLambdaFlowSpec.scala
@@ -2,14 +2,13 @@
  * Copyright (C) 2016-2018 Lightbend Inc. <http://www.lightbend.com>
  */
 
-package akka.stream.alpakka.awslambda.scaladsl
+package docs.scaladsl
 
 import java.util.concurrent.{CompletableFuture, Future}
 
-import scala.concurrent.Await
-import scala.concurrent.duration._
 import akka.actor.ActorSystem
 import akka.stream.ActorMaterializer
+import akka.stream.alpakka.awslambda.scaladsl.AwsLambdaFlow
 import akka.stream.scaladsl.{Keep, Sink}
 import akka.stream.testkit.scaladsl.TestSource
 import akka.testkit.TestKit
@@ -23,6 +22,9 @@ import org.mockito.stubbing.Answer
 import org.scalatest._
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.mockito.MockitoSugar
+
+import scala.concurrent.Await
+import scala.concurrent.duration._
 
 class AwsLambdaFlowSpec
     extends TestKit(ActorSystem("AwsLambdaFlowSpec"))
