+++ docs/src/main/paradox/awslambda.md
@@ -22,18 +22,18 @@ For more information about AWS Lambda please visit the [AWS lambda documentation
 Flow provided by this connector need a prepared `AWSLambdaAsyncClient` to be able to invoke lambda functions.
 
 Scala
-: @@snip [snip](/awslambda/src/test/scala/akka/stream/alpakka/awslambda/scaladsl/Examples.scala) { #init-client }
+: @@snip (/awslambda/src/test/scala/docs/scaladsl/Examples.scala) { #init-client }
 
 Java
-: @@snip [snip](/awslambda/src/test/java/akka/stream/alpakka/awslambda/javadsl/Examples.java) { #init-client }
+: @@snip (/awslambda/src/test/java/docs/javadsl/Examples.java) { #init-client }
 
 We will also need an @scaladoc[ActorSystem](akka.actor.ActorSystem) and an @scaladoc[ActorMaterializer](akka.stream.ActorMaterializer).
 
 Scala
-: @@snip [snip](/awslambda/src/test/scala/akka/stream/alpakka/awslambda/scaladsl/Examples.scala) { #init-mat }
+: @@snip (/awslambda/src/test/scala/docs/scaladsl/Examples.scala) { #init-mat }
 
 Java
-: @@snip [snip](/awslambda/src/test/java/akka/stream/alpakka/awslambda/javadsl/Examples.java) { #init-mat }
+: @@snip (/awslambda/src/test/java/docs/javadsl/Examples.java) { #init-mat }
 
 This is all preparation that we are going to need.
 
@@ -43,10 +43,10 @@ Now we can stream AWS Java SDK Lambda `InvokeRequest` to AWS Lambda functions
 @scaladoc[AwsLambdaFlow](akka.stream.alpakka.awslambda.scaladsl.AwsLambdaFlow$) factory.
 
 Scala
-: @@snip [snip](/awslambda/src/test/scala/akka/stream/alpakka/awslambda/scaladsl/Examples.scala) { #run }
+: @@snip (/awslambda/src/test/scala/docs/scaladsl/Examples.scala) { #run }
 
 Java
-: @@snip [snip](/awslambda/src/test/java/akka/stream/alpakka/awslambda/javadsl/Examples.java) { #run }
+: @@snip (/awslambda/src/test/java/docs/javadsl/Examples.java) { #run }
 
 #### AwsLambdaFlow configuration
